pyproject.toml
[project]
name = "rag-from-scratch"
version = "1.0.0"
description = "An advanced hybrid RAG application"
requires-python = ">=3.12"
dependencies = [
    "pymupdf==1.27.2.2",
    "pydantic-settings==2.12.0",
    "openai==2.32.0",
    "chromadb==1.5.8",
    "elasticsearch==9.2.0",
    "jinja2==3.1.6",
    "httpx==0.28.1",
    "fastapi==0.128.0",
    "ragas==0.4.3",
    "ollama==0.6.1",
    "boto3==1.43.1",
]

[tool.uv]
package = true

[tool.setuptools.packages.find]
include = ["app*", "api*", "eval*"]

[project.scripts]
serve = "api.main:main"
eval = "eval.main:main"
